+ implemented brace matching + replaced surfaces with pixmaps for indent rendering + implemented alpha rectangle rendering * fixed scissors with a hack * changed values, which represents measures to floats * enabled indents and brace highlighting ! probably I missed some values which represent measurements and they still are integers ! fix scissor - it's a hack and should probably be replaced with clipping planes(works only because control uses all the height of the window) ! remove separate alpha rendering - should be supported out of the box with "a" component of color ! refactor: enable alpha blending all the time while control is rendered ! implement lines, ellipses, etc...

+++ src/Style.h
@@ -33,7 +33,7 @@ public:
 	bool hotspot;
 
 	Font font;
-	int sizeZoomed;
+	float sizeZoomed;
 	float lineHeight;
 	float ascent;
 	float descent;
